refactor: restructure access logging for better maintainability
- Reorganize access logging code into separate, focused modules
- Split logger.go into smaller, single-responsibility files
- Extract types and interfaces into dedicated logger_types.go
- Move file writing logic to separate file_writer.go
- Add utility functions in utils.go for common operations
- Update .gitignore to properly exclude log directories

Improves code organization by separating concerns and making
the access logging system more modular and easier to maintain.
